requirement.txt
ta
python-binance
pandas
Clint